Addressed code review comments gradle build passes locally, will check e2e tomorrow
