+++ service/src/main/java/uk/gov/hmcts/reform/fpl/service/CMODocmosisTemplateDataGenerationService.java
@@ -1,0 +1,309 @@
+package uk.gov.hmcts.reform.fpl.service;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.ImmutableMap;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import uk.gov.hmcts.reform.fpl.config.HmctsCourtLookupConfiguration;
+import uk.gov.hmcts.reform.fpl.enums.DirectionAssignee;
+import uk.gov.hmcts.reform.fpl.enums.OtherPartiesDirectionAssignee;
+import uk.gov.hmcts.reform.fpl.enums.ParentsAndRespondentsDirectionAssignee;
+import uk.gov.hmcts.reform.fpl.model.CaseData;
+import uk.gov.hmcts.reform.fpl.model.CaseManagementOrder;
+import uk.gov.hmcts.reform.fpl.model.Direction;
+import uk.gov.hmcts.reform.fpl.model.HearingBooking;
+import uk.gov.hmcts.reform.fpl.model.Respondent;
+import uk.gov.hmcts.reform.fpl.model.Solicitor;
+import uk.gov.hmcts.reform.fpl.model.common.Element;
+import uk.gov.hmcts.reform.fpl.model.common.JudgeAndLegalAdvisor;
+import uk.gov.hmcts.reform.fpl.model.common.Recital;
+import uk.gov.hmcts.reform.fpl.model.common.Schedule;
+import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicList;
+
+import java.io.IOException;
+import java.time.LocalDate;
+import java.time.format.FormatStyle;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+
+import static java.util.Collections.emptyList;
+import static java.util.stream.Collectors.groupingBy;
+import static java.util.stream.Collectors.toList;
+import static org.apache.commons.lang3.ObjectUtils.defaultIfNull;
+import static org.apache.commons.lang3.ObjectUtils.isNotEmpty;
+import static org.apache.commons.lang3.StringUtils.defaultIfBlank;
+import static org.apache.commons.lang3.StringUtils.defaultString;
+import static org.apache.commons.lang3.StringUtils.isBlank;
+import static org.springframework.util.CollectionUtils.isEmpty;
+import static uk.gov.hmcts.reform.fpl.enums.DirectionAssignee.OTHERS;
+import static uk.gov.hmcts.reform.fpl.enums.DirectionAssignee.PARENTS_AND_RESPONDENTS;
+import static uk.gov.hmcts.reform.fpl.service.CaseDataExtractionService.EMPTY_PLACEHOLDER;
+
+// REFACTOR: 02/12/2019 Refactor this with CaseDataExtractionService and NotifyOfProceedingService to try and
+//  extract common elements to CommonCaseDataExtractionService
+@Service
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
+public class CMODocmosisTemplateDataGenerationService extends DocmosisTemplateDataGeneration {
+    private final CommonCaseDataExtractionService commonCaseDataExtractionService;
+    private final CaseDataExtractionService caseDataExtractionService;
+    private final DateFormatterService dateFormatterService;
+    private final DirectionHelperService directionHelperService;
+    private final DraftCMOService draftCMOService;
+    private final HearingBookingService hearingBookingService;
+    private final HmctsCourtLookupConfiguration hmctsCourtLookupConfiguration;
+    private final ObjectMapper mapper;
+
+    @SuppressWarnings("unchecked")
+    public Map<String, Object> getTemplateData(CaseData caseData) throws IOException {
+        ImmutableMap.Builder cmoTemplateData = ImmutableMap.<String, Object>builder();
+        final DynamicList hearingDateList = caseData.getCmoHearingDateList();
+        final String localAuthorityCode = caseData.getCaseLocalAuthority();
+        final CaseManagementOrder caseManagementOrder = defaultIfNull(draftCMOService.prepareCMO(caseData),
+            CaseManagementOrder.builder().build());
+
+        cmoTemplateData.put("familyManCaseNumber", defaultIfNull(caseData.getFamilyManCaseNumber(), EMPTY_PLACEHOLDER));
+        cmoTemplateData.put("generationDate",
+            dateFormatterService.formatLocalDateToString(LocalDate.now(), FormatStyle.LONG));
+        cmoTemplateData.put("complianceDeadline", caseData.getDateSubmitted() != null
+            ? dateFormatterService.formatLocalDateToString(caseData.getDateSubmitted().plusWeeks(26),
+            FormatStyle.LONG) : EMPTY_PLACEHOLDER);
+
+        final List<Map<String, String>> childrenInCase = caseDataExtractionService.getChildrenDetails(caseData);
+        cmoTemplateData.put("children", childrenInCase);
+        cmoTemplateData.put("numberOfChildren", childrenInCase.size());
+
+        cmoTemplateData.put("courtName", getCourtName(localAuthorityCode));
+
+        final String applicantName = caseDataExtractionService.getFirstApplicantName(caseData);
+        cmoTemplateData.put("applicantName", applicantName);
+
+        List<Map<String, String>> respondentsNameAndRelationship = caseDataExtractionService
+            .getRespondentsNameAndRelationship(caseData);
+
+        cmoTemplateData.put("respondents", respondentsNameAndRelationship);
+
+        cmoTemplateData.put("representatives",
+            getRepresentatives(caseData.getRespondents1(), applicantName, caseData.getSolicitor()));
+
+        // Populate with the next hearing booking, currently not captured
+        cmoTemplateData.putAll(commonCaseDataExtractionService.getHearingBookingData(null));
+
+        HearingBooking hearingBooking = hearingBookingService.getHearingBooking(
+            caseData.getHearingDetails(), hearingDateList);
+        JudgeAndLegalAdvisor judgeAndLegalAdvisor = hearingBooking.getJudgeAndLegalAdvisor();
+        cmoTemplateData.putAll(commonCaseDataExtractionService.getJudgeAndLegalAdvisorData(judgeAndLegalAdvisor));
+
+        cmoTemplateData.putAll(getGroupedCMODirections(caseManagementOrder));
+
+        cmoTemplateData.putAll(getDraftWaterMarkData());
+
+        List<Map<String, String>> recitals = buildRecitals(caseManagementOrder.getRecitals());
+        cmoTemplateData.put("recitals", recitals);
+        cmoTemplateData.put("recitalsProvided", isNotEmpty(recitals));
+
+        cmoTemplateData.putAll(getSchedule(caseManagementOrder));
+
+        //defaulting as 1 as we currently do not have impl for multiple CMos
+        cmoTemplateData.put("caseManagementNumber", 1);
+
+        return cmoTemplateData.build();
+    }
+
+    private List<Map<String, Object>> getRepresentatives(List<Element<Respondent>> respondents1,
+                                                         String applicantName,
+                                                         Solicitor solicitor) {
+        List<Map<String, Object>> representatives = new ArrayList<>();
+
+        representatives.add(getApplicantDetails(applicantName, solicitor));
+
+        if (!isEmpty(respondents1)) {
+            respondents1.stream()
+                .map(Element::getValue)
+                .forEach(respondent -> representatives.add(getRepresentative(respondent)));
+        }
+
+        return representatives;
+    }
+
+    private Map<String, Object> getRepresentative(Respondent respondent) {
+        // Currently don't capture respondents representatives apart from solicitor so just return empty placeholders
+        return ImmutableMap.of(
+            "respondentName", respondent.getParty().getFullName(),
+            "representativeName", EMPTY_PLACEHOLDER,
+            "representativeEmail", EMPTY_PLACEHOLDER,
+            "representativePhoneNumber", EMPTY_PLACEHOLDER
+        );
+    }
+
+    private Map<String, Object> getSchedule(CaseManagementOrder caseManagementOrder) {
+        final Schedule schedule = caseManagementOrder.getSchedule();
+        Map<String, Object> scheduleMap = new LinkedHashMap<>();
+
+        if (isScheduleIncluded(schedule)) {
+            scheduleMap.putAll(getEmptyScheduleMap());
+        } else {
+            scheduleMap.putAll(mapper.convertValue(schedule, new TypeReference<>() {}));
+            scheduleMap.put("scheduleProvided", true);
+        }
+
+        return scheduleMap;
+    }
+
+    private boolean isScheduleIncluded(Schedule schedule) {
+        return (schedule == null) || schedule.getIncludeSchedule().equals("No");
+    }
+
+    private Map<String, Object> getEmptyScheduleMap() {
+        final String[] scheduleKeys = {
+            "includeSchedule", "allocation", "application", "todaysHearing", "childrensCurrentArrangement",
+            "timetableForProceedings", "timetableForChildren", "alternativeCarers", "threshold", "keyIssues",
+            "partiesPositions"
+        };
+
+        Map<String, Object> scheduleMap = new LinkedHashMap<>();
+
+        Arrays.stream(scheduleKeys).forEach(key -> scheduleMap.put(key, EMPTY_PLACEHOLDER));
+
+        scheduleMap.put("scheduleProvided", false);
+
+        return scheduleMap;
+    }
+
+    private Map<String, Object> getApplicantDetails(String applicantName, Solicitor solicitor) {
+        Map<String, Object> applicantDetails = new HashMap<>();
+
+        applicantDetails.put("respondentName", defaultIfBlank(applicantName, EMPTY_PLACEHOLDER));
+
+        if (solicitor == null) {
+            applicantDetails.putAll(ImmutableMap.of(
+                "representativeName", EMPTY_PLACEHOLDER,
+                "representativeEmail", EMPTY_PLACEHOLDER,
+                "representativePhoneNumber", EMPTY_PLACEHOLDER
+            ));
+        } else {
+            String phoneNumber = defaultIfBlank(solicitor.getTelephone(), solicitor.getMobile());
+            applicantDetails.put("representativeName", defaultIfBlank(solicitor.getName(), EMPTY_PLACEHOLDER));
+            applicantDetails.put("representativeEmail", defaultIfBlank(solicitor.getEmail(), EMPTY_PLACEHOLDER));
+            applicantDetails.put("representativePhoneNumber", defaultIfBlank(phoneNumber, EMPTY_PLACEHOLDER));
+        }
+
+        return applicantDetails;
+
+    }
+
+    private String getCourtName(String localAuthorityCode) {
+        if (isBlank(localAuthorityCode)) {
+            return EMPTY_PLACEHOLDER;
+        }
+        return defaultIfBlank(hmctsCourtLookupConfiguration.getCourt(localAuthorityCode).getName(), EMPTY_PLACEHOLDER);
+    }
+
+    private Map<String, Object> getGroupedCMODirections(final CaseManagementOrder caseManagementOrder) {
+
+        if (caseManagementOrder == null || isEmpty(caseManagementOrder.getDirections())) {
+            return ImmutableMap.of();
+        }
+
+        Map<DirectionAssignee, List<Element<Direction>>> groupedDirections =
+            directionHelperService.sortDirectionsByAssignee(directionHelperService.numberDirections(
+                caseManagementOrder.getDirections()));
+
+        List<Element<Direction>> parentsAndRespondents = groupedDirections.remove(PARENTS_AND_RESPONDENTS);
+        List<Element<Direction>> otherParties = groupedDirections.remove(OTHERS);
+        ImmutableMap.Builder<String, Object> formattedDirections = ImmutableMap.builder();
+
+        final Map<ParentsAndRespondentsDirectionAssignee, List<Element<Direction>>> groupedParentsAndRespondents =
+            parentsAndRespondents.stream()
+                .collect(groupingBy(directionElement -> directionElement.getValue()
+                    .getParentsAndRespondentsAssignee()));
+
+        final Map<OtherPartiesDirectionAssignee, List<Element<Direction>>> groupedOtherParties = otherParties.stream()
+            .collect(groupingBy(directionElement -> directionElement.getValue().getOtherPartiesAssignee()));
+
+        formattedDirections.put(PARENTS_AND_RESPONDENTS.getValue(),
+            getFormattedParentsAndRespondentsDirections(groupedParentsAndRespondents));
+
+        formattedDirections.put(OTHERS.getValue(), getFormattedOtherPartiesDirections(groupedOtherParties));
+
+        groupedDirections.forEach((key, value) -> {
+            List<Map<String, String>> directionsList = buildFormattedDirectionList(value);
+            formattedDirections.put(key.getValue(), directionsList);
+        });
+
+        return formattedDirections.build();
+    }
+
+    private List<Map<String, Object>> getFormattedOtherPartiesDirections(
+        Map<OtherPartiesDirectionAssignee, List<Element<Direction>>> groupedOtherParties) {
+
+        List<Map<String, Object>> directionsToRespondents = new ArrayList<>();
+        groupedOtherParties.forEach((key, value) -> {
+            Map<String, Object> directionForRespondent = new HashMap<>();
+            directionForRespondent.put("header", "For " + key.getLabel());
+            List<Map<String, String>> directionsList = buildFormattedDirectionList(
+                value);
+            directionForRespondent.put("directions", directionsList);
+            directionsToRespondents.add(directionForRespondent);
+        });
+
+        return directionsToRespondents;
+    }
+
+
+    private List<Map<String, Object>> getFormattedParentsAndRespondentsDirections(
+        Map<ParentsAndRespondentsDirectionAssignee, List<Element<Direction>>> groupedParentsAndRespondents) {
+
+        List<Map<String, Object>> directionsToRespondents = new ArrayList<>();
+        groupedParentsAndRespondents.forEach((key, value) -> {
+            Map<String, Object> directionForRespondent = new HashMap<>();
+            directionForRespondent.put("header", "For " + key.getLabel());
+            List<Map<String, String>> directionsList = buildFormattedDirectionList(
+                value);
+            directionForRespondent.put("directions", directionsList);
+            directionsToRespondents.add(directionForRespondent);
+        });
+
+        return directionsToRespondents;
+    }
+
+    private List<Map<String, String>> buildFormattedDirectionList(List<Element<Direction>> directions) {
+        return directions.stream()
+            .map(Element::getValue)
+            .filter(direction -> !"No".equals(direction.getDirectionNeeded()))
+            .map(direction -> ImmutableMap.of(
+                "title", formatTitle(direction),
+                "body", defaultIfNull(direction.getDirectionText(), EMPTY_PLACEHOLDER)))
+            .collect(toList());
+    }
+
+    private String formatTitle(Direction direction) {
+        return String.format("%s by %s",
+            direction.getDirectionType(),
+            direction.getDateToBeCompletedBy() != null
+                ? dateFormatterService.formatLocalDateTimeBaseUsingFormat(
+                direction.getDateToBeCompletedBy(), "h:mma, d MMMM yyyy")
+                : "unknown");
+    }
+
+    private List<Map<String, String>> buildRecitals(final List<Element<Recital>> recitals) {
+        if (isEmpty(recitals)) {
+            return emptyList();
+        }
+
+        return recitals.stream()
+            .filter(Objects::nonNull)
+            .map(Element::getValue)
+            .map(recital -> ImmutableMap.of(
+                "title", defaultString(recital.getTitle(), EMPTY_PLACEHOLDER),
+                "body", defaultString(recital.getDescription(), EMPTY_PLACEHOLDER)
+            ))
+            .collect(toList());
+    }
+}
