+++ service/src/test/java/uk/gov/hmcts/reform/fpl/service/HearingBookingServiceTest.java
@@ -1,22 +1,33 @@
 package uk.gov.hmcts.reform.fpl.service;
 
 import com.google.common.collect.ImmutableList;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.hmcts.reform.fpl.model.CaseData;
 import uk.gov.hmcts.reform.fpl.model.HearingBooking;
 import uk.gov.hmcts.reform.fpl.model.common.Element;
+import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicList;
+import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicListElement;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
+import static java.util.UUID.fromString;
 import static org.assertj.core.api.Assertions.assertThat;
 import static uk.gov.hmcts.reform.fpl.utils.CaseDataGeneratorHelper.createHearingBooking;
 
 @ExtendWith(SpringExtension.class)
 class HearingBookingServiceTest {
+
+    private static final UUID[] UUIDS = {
+        fromString("b15eb00f-e151-47f2-8e5f-374cc6fc2657"),
+        fromString("6b3ee98f-acff-4b64-bb00-cc3db02a24b2"),
+        fromString("ecac3668-8fa6-4ba0-8894-2114601a3e31")
+    };
 
     private final HearingBookingService service = new HearingBookingService();
     private static final LocalDateTime TODAYS_DATE = LocalDateTime.now();
@@ -55,18 +66,77 @@ class HearingBookingServiceTest {
         assertThat(sortedHearingBooking.getStartDate()).isEqualTo(TODAYS_DATE);
     }
 
+    @Nested
+    class GetHearingBooking {
+        private DynamicList dynamicList;
+        private List<Element<HearingBooking>> hearingDetails;
+
+        @BeforeEach
+        void setUp() {
+            dynamicList = createDynamicList();
+            hearingDetails = createHearingBookings();
+        }
+
+        @Test
+        void shouldReturnAEmptyHearingBookingWhenHearingDetailsIsNull() {
+            final HearingBooking hearingBooking = service.getHearingBooking(null, dynamicList);
+            assertThat(hearingBooking).isEqualTo(HearingBooking.builder().build());
+        }
+
+        @Test
+        void shouldReturnAEmptyHearingBookingWhenTheDynamicListIsNull() {
+            final HearingBooking hearingBooking = service.getHearingBooking(hearingDetails, null);
+            assertThat(hearingBooking).isEqualTo(HearingBooking.builder().build());
+        }
+
+        @Test
+        void shouldReturnAEmptyHearingBookingWhenTheDynamicListValuesCodeIsNull() {
+            dynamicList = DynamicList.builder().build();
+            final HearingBooking hearingBooking = service.getHearingBooking(hearingDetails, dynamicList);
+            assertThat(hearingBooking).isEqualTo(HearingBooking.builder().build());
+        }
+
+        @Test
+        void shouldReturnAEmptyHearingBookingWhenHearingDetailsIsEmpty() {
+            final HearingBooking hearingBooking = service.getHearingBooking(List.of(), dynamicList);
+            assertThat(hearingBooking).isEqualTo(HearingBooking.builder().build());
+        }
+
+        @Test
+        void shouldReturnTheFirstHearingBookingWhenTheDynamicListValueCodeMatches() {
+            final HearingBooking hearingBooking = service.getHearingBooking(hearingDetails, dynamicList);
+            assertThat(hearingBooking).isEqualTo(hearingDetails.get(0).getValue());
+        }
+
+        private DynamicList createDynamicList() {
+            return DynamicList.builder()
+                .value(createDynamicElement(UUIDS[0]))
+                .listItems(List.of(
+                    createDynamicElement(UUIDS[0]),
+                    createDynamicElement(UUIDS[1]),
+                    createDynamicElement(UUIDS[2])
+                ))
+                .build();
+        }
+
+        private DynamicListElement createDynamicElement(UUID code) {
+            return DynamicListElement.builder().code(code).label("").build();
+        }
+
+    }
+
     private List<Element<HearingBooking>> createHearingBookings() {
         return ImmutableList.of(
             Element.<HearingBooking>builder()
-                .id(UUID.randomUUID())
+                .id(UUIDS[0])
                 .value(createHearingBooking(TODAYS_DATE.plusDays(5), TODAYS_DATE.plusDays(6)))
                 .build(),
             Element.<HearingBooking>builder()
-                .id(UUID.randomUUID())
+                .id(UUIDS[1])
                 .value(createHearingBooking(TODAYS_DATE.plusDays(2), TODAYS_DATE.plusDays(3)))
                 .build(),
             Element.<HearingBooking>builder()
-                .id(UUID.randomUUID())
+                .id(UUIDS[2])
                 .value(createHearingBooking(TODAYS_DATE, TODAYS_DATE.plusDays(1)))
                 .build()
         );
